requirements.txt
django
djangorestframework
python-dotenv
requests
django-storages
boto3
accelerate>=0.26.0 